refactor components package
