+++ src/main/scala/it/unibo/authsim/client/app/components/persistence/UserEntity.scala
@@ -1,4 +1,4 @@
-package it.unibo.authsim.components.persistence
+package it.unibo.authsim.client.app.components.persistence
 
 case class UserEntity(username: String, password: String)
